% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dtdt.R
\name{calc_dtdt}
\alias{calc_dtdt}
\title{Transmission-Disequilibrium Test Statistic}
\usage{
calc_dtdt(Y, trio_geno, is_phased = FALSE)
}
\arguments{
\item{Y}{The n-length vector of the children's dichotomous trait}

\item{trio_geno}{A \eqn{3 \times n}{3 x n} matrix of trio genotype if `FALSE`
for is_phased. Or a \eqn{4 \times n}{4 x n} matrix of trio genotype if
`TRUE` for is_phased. Must contain only values \eqn{(0, 1, 2)}{0, 1, 2}.
The first row is assumed maternal, the second row is assumed paternal, and
the third (or together with the fourth) row is assumed child's genotype.}

\item{is_phased}{Label for unphased/phased genotype. If `FALSE` (default), the
child's genotype is unphased. Else, the child's genotype is phased.}
}
\value{
A numeric value of the TDT statistic.
}
\description{
This function calculates the TDT statistic given trio genotypes and children's 
dichotomous phenotype.
}
\note{
Inputting phenotypes that are not dichotomous and undesired dimensions
    of the genotype matrices will result in errors.
}
\examples{
trio_example <- matrix(c(1,1,2,0, 1,2,1,0, 1,1,1,0), nrow=3, byrow=TRUE) 
y_binary <- c(0,1,1,0)
dtdt <- calc_dtdt(y_binary, trio_example)

}
