% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dtmt_pc.R
\name{calc_dtmt_pc}
\alias{calc_dtmt_pc}
\title{This function computes the centered TMT statistic given trio genotypes and 
children's phenotype. The phenotype can be either quantitative or dichotomous.}
\usage{
calc_dtmt_pc(Y, trio_geno, mu, is_phased = FALSE)
}
\arguments{
\item{Y}{The n-length vector of children' phenotype.}

\item{trio_geno}{A \eqn{3 \times n}{3 x n} matrix of trio genotype if `FALSE`
for is_phased. Or a \eqn{4 \times n}{4 x n} matrix of trio genotype if
`TRUE` for is_phased. Must contain only values \eqn{(0, 1, 2)}{0, 1, 2}.
The first row is assumed maternal, the second row is assumed paternal, and
the third (or together with the fourth) row is assumed child's genotypes.}

\item{mu}{A numeric value of the population mean.}

\item{is_phased}{Label for unphased/phased genotype. If `FALSE` (default), the
child's genotype is unphased. Else, the child's genotype is phased.}
}
\value{
A numeric value of the centered TMT statistic.
}
\description{
This function computes the centered TMT statistic given trio genotypes and 
children's phenotype. The phenotype can be either quantitative or dichotomous.
}
\note{
Undesired dimensions of the genotype matrices will result in errors.
}
\examples{
trio_example <- matrix(c(1,1,2,0, 1,2,1,0, 1,1,1,0), nrow=3, byrow=TRUE) 
list_assign <- assignment_index(trio_example, is_phased=FALSE)
y_quant <- c(11.8,1.7,3.6,12.2)
dtmt_pc <- calc_dtmt_pc(y_quant, trio_example, mu=5)

}
