% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mu_hat.R
\name{calc_mu_hat}
\alias{calc_mu_hat}
\title{Population Mean Estimate}
\usage{
calc_mu_hat(Y, W0, W1)
}
\arguments{
\item{Y}{The n-length vector of children' phenotype.}

\item{W0}{A vector of indicators for assigning to group 0.}

\item{W1}{A vector of indicators for assigning to group 1.}
}
\value{
A numeric value of the population mean estimate.
}
\description{
This function computes the unbiased estimate for the population mean.
}
\examples{
trio_example <- matrix(c(1,1,2,0, 1,2,1,0, 1,1,1,0), nrow=3, byrow=TRUE) 
y_quant <- c(11.8,1.7,3.6,12.2)
list_assign <- assignment_index(trio_example)
mu_hat <- calc_mu_hat(y_quant, list_assign$W0, list_assign$W1)

}
