% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{familydata}
\alias{familydata}
\title{Example Nuclear Family Data.}
\format{
A list of matrices:
\describe{
  \item{pheno_offspring}{Offspring Phenotype}
  \item{geno_offspring}{Offspring Genotypes}
  \item{geno_maternal}{Maternal Genotypes}
  \item{geno_paternal}{Paternal Genotypes}
  \item{fam_id_maternal}{Maternal Family ID}
  \item{fam_id_paternal}{Paternal Family ID}
  \item{fam_id_offspring}{Offspring Family ID}
}
}
\source{
Simulated data
}
\usage{
familydata
}
\description{
Example Nuclear Family Data.
}
\keyword{datasets}
