% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{triolist}
\alias{triolist}
\title{Example Trio Data}
\format{
A list of matrices:
\describe{
 \item{geno_m}{A matrix of the maternal genotypes.}
 \item{geno_p}{A matrix of the paternal genotypes.}
 \item{geno_c}{A matrix of the child's genotypes}
 \item{allele_m}{A matrix of the maternal transmitted alleles.}
 \item{allele_p}{A matrix of the paternal transmitted alleles.}
 \item{y}{A single row matrix of the child's phenotypes.}
}
}
\description{
Example Trio Data
}
\examples{
\dontrun{
 names(triolist)
}
}
